Remove BigDigClient logic from SshHandshake
Summary:
While it's convenient to have `SshHandshake` take care of establishing the client too, it's not ideal for the way Nuclide uses `SshHandshake`. (Nuclide's `SshHandshake` is also wired to return a `RemoteConnection` afterward, which is also nasty).

To handle connections in a more implementation-agnostic way it ends up being cleaner if we return the necessary parameters for `createBigDigClient` and let the caller do it instead. This allows us to share a common code path to reconnect and establish `RemoteConnection`s within Nuclide rather than having two.

It seems that `fb-vscode/src/RemoteConnect.js` is already working outside of `SshHandshake` to reconnect to an existing BigDig server which does indicate that this is a good path forward.

Finally, it seems to make sense from a conceptual standpoint that `SshHandshake` does the bare minimum of obtaining credentials via SSH and then the caller should do the rest.

Reviewed By: siegebell

Differential Revision: D6855677

fbshipit-source-id: 93ae33026caf0bf795b512ad8c5e0be70151418b