Plug in big-dig's SshHandshake into Nuclide's connection dialog
Summary:
Add an alternate code path that uses big-dig's SSH handshake when the `nuclide_big_dig` GK passes.
I tried to be as minimally disruptive as possible.

There's some slight cleanup in `ConnectionDialog` to represent the current `SshHandshake` state as an Rx subscription to prevent races with async code and handshake cancellation.

Reviewed By: semmypurewal

Differential Revision: D6809732

fbshipit-source-id: 633b7246d05178f2dc9dd5dca88a4fbd9ae721d5