Clean up RemoteConnection reconnnect logic
Summary:
This is shared across the old `SshHandshake` code and the `nuclide-remote-projects` startup path but the code is duplicated.
Let's clean that up into a single public API (`RemoteConnection.reconnect`). `_createConnectionBySavedConfig` isn't really what you want anyway.

The slight cost here is that this ends up doing the DNS lookup twice.. but I don't think that should be significant at all with the OS-level DNS cache.

There also appears to be some legacy code in the startup path that isn't necessary anymore (`createConnectionBySavedConfig` should never return something with a CWD that doesn't match.)

Reviewed By: semmypurewal

Differential Revision: D6856306

fbshipit-source-id: 0a149119bd98356c6b1fd753a64598571941b249