Analytics tracking for all running processes
Summary:
This adds tracking to fire a `process-exit` analytics event whenever a process spawned by Nuclide exits.

As a preface, the complexity of this is several-fold:

1) `process` is in `nuclide-commons`, so it doesn't have reliable access to `nuclide-analytics`.
2) So the first solution would be to export a `Subject` from `process`. But it's also running from several contexts: Atom, local RPC, and remote servers.

It seems pretty gross to add something to every potential entry point, but we already have one in the form of `iniitalizeLogging`. And we're also already logging the results of all processes, so combine the two together and..

The solution here is to log a custom `ProcessLoggingEvent` to `log4js` instead of logging a string (as we currently do).

- `ProcessLoggingEvent` overrides `[util.inspect.custom]()` so it still gets logged as a string.
- We add a custom appender `processTrackingAppender` that subscribes to the `nuclide-commons/process` category and intercepts these events.
- We can grab the metadata from `ProcessLoggingEvent` and then track it - but this is a very high volume so it has to be sampled at a 1/10 rate. Commands may also be very long so they're truncated to 100 characters.

I noticed two minor fixes while writing this:

1) `logError` can use `logger.error` instead of `console` (which goes nowhere.)
2) It's nice to shell quote the full commands - that way they can be copy-pasted into a shell for manual testing.

Reviewed By: matthewwithanm

Differential Revision: D6868738

fbshipit-source-id: 6bd6319393c0cacb1647155891fff446a19bd4bf