Fix redbox from quick-open
Summary:
I got a redbox after clicking a result in quick-open in Atom 1.24.

Seems to be a pretty clear race between the `focus()` call and the quick-open dialog closing itself after clicking a result.
The repro isn't clear, but the easy fix is to properly flowtype `getTextEditorElement()` to be nullable.

Reviewed By: aadisriram

Differential Revision: D6891176

fbshipit-source-id: 7bc1e4ffd58924b5d18a27696fa51a95508ab85b