Avoid "not supported" warnings from Flow find-refs
Summary:
Returning 'null' from a find-refs request indicates that the provider doesn't apply (and a warning message appears with "find-refs is not supported for the current project".

There's a separate return type for errors. It looks like there are two clear cases where an error is more appropriate.

(Find-refs errors are also tracked so we can see how many users are actually afflicted by each of these in the wild.)

Reviewed By: nmote

Differential Revision: D6897050

fbshipit-source-id: d9dc868557c8fc740706d5047c9862a40df61b35