Exclude trailing '.' when opening urls in terminal
Summary:
It is pretty common for terminal programs to emit a link, and often it ends a sentence. Because of the spec for URLs, the '.' that ends the sentence is technically part of the URL. However, while many web servers remove trailing '.' characters, many do not. This does the wrong thing in the common case. Many terminal programs do not include a trailing '.'.

Although useful, because this violates the URL spec, it seems more of a matter of opinion rather than something to push upstream to xterm.js. This means the hover decoration does include the '.'.

Reviewed By: pelmers

Differential Revision: D6934741

fbshipit-source-id: b27c08ea20cbbdd88537ce9655fc24b876effb28