avoid generating unused analytics
Summary:
In nuclide-health, we calculate app-health information around DOM counters, process statistics on sub-processes started by Nuclide, event-loop stalls, and memory usage for our Atom process. In the open-source build, none of this information is tracked because the 'track' method where all of this ends up is empty.

However, while investigating https://github.com/facebook/nuclide/issues/1455, I realized that Nuclide still calculates this information before discarding it.

The issue in 1455 is that the subscription does not correctly handle the error path.

This change adds an error handler that just logs. It also avoids calculating any of this data in the first place in the open-source build since it will be discarded anyway.

Reviewed By: hansonw

Differential Revision: D6934559

fbshipit-source-id: c7e4f702ecbc27f45f2608bf9617911e881895d1