improve buck header handling
Summary:
there are some targets, like the one for fbobjc/xplat/effects-framework/scene/rendertarget/DoodlingProcessor.hpp which are specifically related to headers and have empty compilation databases, which renders the language service unusable.
Given that our header -> source finder is getting better, we can use it for handling headers in the buck provider.
In short, now we always get the compilation flags from the corresponding source of a given header.

Reviewed By: pelmers

Differential Revision: D6933900

fbshipit-source-id: 86ce947b32c5f12dec82a57519f1c25915da6893