Fix using default flags
Summary:
My previous ClangFlags refactor made the flags argument non-optional, so it could not be null. Instead the check should be that the flags have zero length, see ClangFlagsManager#_getFlagsForSrcImpl.

At the time we didn't see this error remotely because of another error in calling the RelatedFileFinder in the previous block of ClangFlagsManager always throwing a host mismatch, which D6879040 fixed and uncovered this one.

Reviewed By: a20012251

Differential Revision: D6961476

fbshipit-source-id: c79033dd839b64090ea997fea7269ee02dd35ddc