attempt to fix bug and gather more data about it
Summary:
Attempt to find the bug in the socket service that's crashing the server.
The hypothesis is that in an error condition, we end the socket and the server is
still sending data after that (which means we've already disposed the remote connection).
I wasn't able to repro this in a test, and after carefully going through the code, I'm
not 100% convinced that this is the problem. So I'm adding some state to track when we should
no longer be writing to the remote connection, and using that as a guard. I'm also sending
data to the server when we hit that condition (via the nuclide-analytics package and log4js's
error).

Reviewed By: stepanhruda

Differential Revision: D6943233

fbshipit-source-id: c73b9628bb783e36c191053561e830f1e2fd3f8e