Ocaml 'logLevel' now user-configurable
Summary:
Previously the Ocaml log-level was set to 'INFO'.

That was a bit of a pain for LSP-related traces, which come at the lower 'TRACE' level and so weren't being shown.

This diff makes Ocaml log-level a user-configurable setting under View > Toggle Nuclide Settings, similar to how Hack does it.

I've changed the default to be 'DEBUG'. If figure that's appropriate for most end-users. We LSP/OCaml developers can set it to 'ALL' for our work.

Reviewed By: pittsw

Differential Revision: D6953985

fbshipit-source-id: 6eb91307dc35c7b06dce230c789e3672a57d4568