Add function to get all the filenames from a db
Summary: The buck plugin used to request the entire mapping of all files to their flags only to throw away the flags immediately. Add a new function to the clang service to only get the filenames allowing us to remove the limit on the buck cache size (since we only store filenames).

Reviewed By: a20012251

Differential Revision: D6954731

fbshipit-source-id: 1127a81aa5359e2d417e1d9dc2a0b22ca4b9865b