Change diff parsing to be able to handle file add/move/deletion cases
Summary:
Diff parsing was not handling unified diffs that represented files being moved or added, as the first line of an add patch is /dev/null, and a move/copy patch is the name of the old file. This change makes the parser try to use the name of the new file first, and if it was /dev/null, representing a deletion, it would check the name of the old file.
This should be strictly more correct than the old version.

Reviewed By: aadisriram

Differential Revision: D6973188

fbshipit-source-id: 90bfdf3cfdd098350a1a9d8bdfa7bea0c06e9a2c