Fix diff computation line mapping to map lines at the end of the file correctly
Summary: Previously, we have had the diff parser return a line mapping that had the last two entries incorrect. We would expect the mapped line sequence to monotonically increase, whereas in the case where the "before" file was shorter than the "after" file, we would get the last two lines of the oldToNew mapping be the number of the last 2 lines of the old file.

Reviewed By: aadisriram

Differential Revision: D6973195

fbshipit-source-id: d5f523f241e5883da037c03a432140b3ddadd7ec