Limit file events by project
Summary:
Previously all file events for all c++ files were sent to all the cquery processes running which meant that all of them except the correct one would emit bad diagnostics information.
Now we store all known filenames when creating the process and make sure the edited file matches. It's not perfect yet because it does not remember header files, but it's a start.

Reviewed By: a20012251

Differential Revision: D6951266

fbshipit-source-id: 0ddea4b5bfcedfc0af34daa30aea855e04a886a4