Dont Request Unused Password When Creating A New Connection Profile
Summary:
Before this diff, if you create a new profile, then you need to enter your password. But then Nuclide says for security reasons it won't save the password. It doesn't even use the password. And then the next time along, when you decide to connect to the remote project using this profile, you need to reenter the password.

This diff makes it so you don't need to do that.

The second image in the test plan is the one that should and does look different than before.

Reviewed By: wbinnssmith

Differential Revision: D6956243

fbshipit-source-id: 9d40c23bcbbc17a4666abd4f43b5443353994c0e