Nuclide debugger: fix race between resume and source line selection
Summary:
There's a race caused by the VSP translator selecting a call frame's source line in the editor and a resume happening, especially at the loader breakpoint. By the time we dispatch the call to open an editor and highlight a line, the debugger has already resumed. This is confusing users - they think the debugger has paused because a line is highlighted in the editor, but the controls, stack frame and threads panes show the debugger running.

This diff has us not select anything in the text editor if the debugger is not paused any more.

Reviewed By: mostafaeweda

Differential Revision: D6983531

fbshipit-source-id: 8f9cd49a40b2bd925086fc1509d9f902764f673c