wire up NuclideSocket in big dig, no reconnection (yet)
Summary:
Use NuclideSocket in BigDig. There's still something wrong with this. When you kill
the Nuclide Server and then connect, it seems the heartbeat failure is not registered anywhere
so Nuclide waits for a timeout from nuclide rpc before showing the yubikey/duo dialog. When
we finally connect, however, everything seems to work so I'll fix the heartbeat issue under
separate cover.

The next diff in the stack wires up the reconnection when the network goes out.

Reviewed By: hansonw

Differential Revision: D7014371

fbshipit-source-id: 412faf88986d64d9c0f5c725997eac40679f0b45