Allow starting server with existing certs
Summary:
1. Add new arguments to `nuclide-start-server` with paths to certificates to reuse instead of generating new ones.
2. When starting a server that generates certificates (without these new args), we print out the paths with the rest of the stdout (it already contains the client certs).

 ---

Q: Why print paths and not the certificate contents themselves?
A: Paths get passed as args to the nuclide-main node process. If we passed the contents, you could easily see the private key if you `ps aux | grep nuclide-main`. Teems more conservative/less invasive (maybe you don't have read permissions to that path)? It does mean there is currently a discrepancy because stdout prints contents for client certs (not passed to server), but only paths for server certs.

Q: the package has an implicit dependency on client cert files existing next to server cert files so it can print client certs to stdout
A: It seemed fine for now. At the moment we only use the generator provided by this same package so they will always be there.

Reviewed By: hansonw

Differential Revision: D6993610

fbshipit-source-id: 0eaa0383bfcb59a6f3e376fb11eb4827ddd1b555