Fail tests if there are Flow warnings
Summary:
Unused suppressions are warnings now, because of lazy mode. Still, we're doing a full check in CI anyway and we don't want to let them through. D7070403 reminded me of this.

We can use the new `max-warnings` flag which I added in Flow v0.66 to make sure we're not letting suppressions through.

Reviewed By: wbinnssmith

Differential Revision: D7071606

fbshipit-source-id: 5efe610571aeb7073a5b6a9eed2420efd9ca067d