Migrate HHVM debuggers to VspProcessInfo
Summary:
Moves the HHVM debuggers (both the old XDebug and new VS Debug) off of their own process info objects and onto objects that extends VspProcessInfo. XDebug and VSP debuggers are both supported for now, and we'll remove the XDebug code when it's no longer supported.

Support for "run in terminal" for the HHVM VSP debugger is maintained, but support for running XDebug in the terminal has been cut.

Reviewed By: ebluestein

Differential Revision: D7012284

fbshipit-source-id: 40d0785c06c7b06561a70281f76ebc4826ba206d