fallback to any source in the same folder when finding sources for headers in buck projects
Summary:
According to yiding jia and my readings, the includes are the same for all the files in the same compilation database and the compilation flags are almost always the same (some files might have additional flags, but that's not an issue), which means that, when finding a corresponding source for a header that belongs to a buck project, using any source in the same folder is a good fallback.

In order to quickly determine is the file belongs to a buck project, we just look for a BUCK or TARGETS file close in the tree. Asking buck for the owner of a source file in the same folder might also work, but it takes several seconds and probably is not necessary.

Reviewed By: pelmers

Differential Revision: D7013075

fbshipit-source-id: 7b0102658bbbb59edbfa113d1bdb8b614fda7bd0