Use git format for hg show
Summary: the hg diff format is hacky and badly supported according to the source control team. Using --git leads to better naming and supports empty files, etc. The reason it's not the default is because it can apparently be slower. I haven't noticed this in my testing, but if this ends up being a perf regression we can revert easily given its such a small change.

Reviewed By: tjfryan

Differential Revision: D7102001

fbshipit-source-id: f916d4c55b5c37cb1de89307820a020c4be0bc43