Limit the number of fetched statuses from hg
Summary: In some circumstances, users may be in a weird state and hg will try to return way more changed files than it should. Even if there are thousands of files changed in a valid way, we never need to show the user 1000 things. We can limit this to a sane number to help prevent UI freezes

Reviewed By: tjfryan

Differential Revision: D7106462

fbshipit-source-id: 7d0904218e330025b081e36634e949319e2aecb4