HHVM Debugger: dummy console request needs to target correct sandbox
Summary: The request that backs the console REPL while the target is running needs to be affinitized to the current user's sandbox or else it will pull documents from the server's default fallback PHP ROOT instead of the sandbox's. This will cause modified code not to be observed, and breakpoints to be missed for routines invoked from the console.

Reviewed By: velocityboy

Differential Revision: D7106309

fbshipit-source-id: 08e57adec9242e99e6d8295d93b10b57e7c72b89