Make parser colon-robust and relax invariants
Summary:
The search tools output being colon separated causes issues if either the filename or the line contains colons.

rg provides a '-0' option and grep provides '--null' which will delimit the filename with a null byte.

However because the other tools ('hg/git grep', 'ack') do not provide this option we should relax the match invariant and skip the result instead of throwing.

Reviewed By: a20012251

Differential Revision: D7117886

fbshipit-source-id: 3b5338e6bd218938409fefebd200b267ad9cce5f