Disable Clicking Stackframes if Sources are not Available
Summary:
If a row in the callstack doesn't have an associated source (frame.source.available is false), then do not allow it to be clickable and additionally dim that row so that it is clear it is disabled.

The dim is a very subtle effect on just the text of the row.

Reviewed By: mostafaeweda

Differential Revision: D7131270

fbshipit-source-id: 75c16d5b7500ccff96bd05421c3f25cd870db87c