Tag nuclide-task processes with a name
Summary:
When looking at bug reports, we often see `bootstrap.js` processes consuming CPU/memory (from the `nuclide-task` library).

Unfortunately, it's impossible to tell which of the `nuclide-task` callsites is responsible from just the process command (there are a couple now). This adds a mandatory `name` field that gets added to the command line (as a no-op) for tagging purposes.

Reviewed By: mostafaeweda

Differential Revision: D7132601

fbshipit-source-id: cb2aa306731e14419deed3d363c8ffedccbc57c2