HHVM Debugger: indicate which stop thread should have focus in the client UX
Summary:
This adds an experimental threadCausedFocus attribute to a stopped event so the client UX can determine which thread should be focused for the user when multiple threads have stopped. This attribute is currently proposed as an addition to the VS Code debugging protocol, but not yet implemented in trunk.

Tracked by GitHub issue: https://github.com/Microsoft/vscode-debugadapter-node/issues/147

Reviewed By: mostafaeweda

Differential Revision: D7141866

fbshipit-source-id: 2495ee5e742470f1c7af3a3017db42e093be3db0