Remove PanelRenderer
Summary:
It wasn't paying the abstraction tax so I killed it.

While I was in there, I also addressed a few race conditions caused by Promises/async. For example, if the breadcrumbs package was deactivated after the GK check completed, the panel could never be disposed. Similarly, actions could be dispatched on the store after the package was disposed.

Reviewed By: captbaritone

Differential Revision: D7207366

fbshipit-source-id: 162c845c766d5a8638e211bdf1d4e1ec0fc67ee2