Allow multi-line input in debugger console
Summary:
This diff adds a tip telling the user they can Shift+Enter (Windows/Linux) or Option+Enter (Mac) to enter multi-line text in the debugger console prompt.

When pressing the up or down arrows, the input area will suggest history auto complete only if the input has 1 or fewer lines in it, so we don't replace your input with history when you're navigating around a multi-line text area.

This diff also fixes an issue where really long unbreakable text in a console record is truncated because there was no scrollbar.

Reviewed By: captbaritone

Differential Revision: D6861199

fbshipit-source-id: 305e810b3f9ff38d6f1aae1d44d4557e767e3551