Set initial state before opening a tunnel
Summary:
Sometimes a tunnel could get open very quickly and we'd try to update its state before we had even marked it as `initializing`. The `SET_TUNNEL_STATE` action assumes that we have some kind of state already and would crash on this:

```
invariant(state.get(action.payload.tunnel) != null);
```

Getting rid of the invariant isn't a good idea, because we would update the state to `ready` and immediately overwrite it back to `initializing`. This makes sure the sequence of events happens as intended.

Reviewed By: semmypurewal

Differential Revision: D7221298

fbshipit-source-id: 80f1f12f28f84f123c80ef96e1db35f78a7d776b