Improve hg progress status component
Summary:
I ran into a situation where hg did some discovering in the background but got stuck. I believe this is due to a race where the progress file changes faster than watchman debouncing and we only get one event and never clear the state.

As a precaution, we should only show progress messages while an action is actually going on.

Furthermore, progress messages actually have an "active" field which we should respect, since it causes it to show more relevant messages (in my testing).

Reviewed By: tjfryan

Differential Revision: D7222440

fbshipit-source-id: 1aaccd71d95e06d7f78c5774f82c505cb5e29b2b