HHVM Debugger: Support terminate request command
Summary:
This adds support for a new "terminateThread" command that will kill the specified thread by raising a PHP fatal exception on it. This can be useful during  debugging for preventing a thread from continuing down a particular code path, or aborting before side effects are produced when debugging an issue.

terminateThread is not yet available in the VS Code debugger protocol spec, so it is prefixed with "fb_" to denote it is Facebook-specific.

Reviewed By: velocityboy

Differential Revision: D6850122

fbshipit-source-id: e22150d1e0aa682baed55de9f15661235cf591fc