allow nuclide server to accept clientIds in the initial ws connection
Summary:
D7053394 simplifies NuclideSocket by sending the clientId in the initial websocket connection. Unfortunately this causes our upgrade path to fail in the BigDig case, since the old version of the server will be expecting the clientId to be sent as the first message, but the client will be sending a different message at that point. This diff gives us a path to upgrade without impacting users. So the plan is to roll this diff out to production, then land the clientId diff, then delete this additional code.

It seems that the old NuclideServer path won't be affected since the heartbeat will be sent first and the version check will happen. Unfortunately, I couldn't confirm that path without a release build, so I went ahead and did the upgrade path there as well.

Reviewed By: hansonw

Differential Revision: D7258076

fbshipit-source-id: 19d7b2c6f1de075745707b3902eea8745736bada