Add nuclide atomproject
Summary:
Add a Nuclide project file to JSON. This is landing before the rest of project configs is out because I would like to have a *.atomproject.{cson, json} file in fb to test my atomproject reading job (aka: php script scanning for atomproject files).

I added paths in (even though technically they are redundant) because in general I think it is better to be explicit about which paths a project launches.

@[100002839212206:matthewwithanm] I included all of the feature groups by default, since if a feature starts erroring, we should probably know about it right away. If anyone feels strongly otherwise, please let me know!

Reviewed By: hansonw

Differential Revision: D7274575

fbshipit-source-id: 8a9b95f12c18c1dec7c02183cc09877034145310