Rework tunneling implementation
Summary:
This now supports tunneling to multiple hosts concurrently and proper cleanup. I originally implemented a more sophisticated solution in mind which handled multiple subscribers, but ended up going back to this.

The assumption is that tunneling is either on or off, and it should be okay for a feature to turn it on and leave it on, instead of triggering it there and back.

Reviewed By: aadsm

Differential Revision: D7282718

fbshipit-source-id: d9e7e6d03be770805c4ee11705c6e0f16ae43212