Create Atom reporter
Summary:
This creates an Atom reporter for running Jest suites.

Note that this does not use Less and uses inline styles. This is because we run outside the typical Atom context, meaning we would have to bring the less runtime with us. This was easier and allowed me to keep the frontend concise and in one file.

Reviewed By: matthewwithanm

Differential Revision: D7215900

fbshipit-source-id: 16480f882cc1a3f2c008ff664513dda7206113e1