Make code lens not scroll a newly opened file
Summary: When opening a file to a line other than 1 (such as through a BigGrep search in the OmniSearch box), code lens has a tendency to scroll the file a few ms after it's opened because it's adding the dummy lines.  This change will recenter the file after the next tick of the event loop.

Reviewed By: ljw1004

Differential Revision: D7296013

fbshipit-source-id: 5d01a22c8a1d9a8955a8a68ac6a4689d27811d1b