Don\'t store a failed open tunnel in the cache
Summary:
Deletes the promise entry from the cache whenever it gets rejected.
I do this because otherwise it's not possible to retry to create the tunnel unless we reload Atom.

I'm doing the same thing for when the tunnels get disposed because otherwise we end up with promises in the cache for tunnels that have already been disposed.

Reviewed By: stepanhruda

Differential Revision: D7301350

fbshipit-source-id: f4276c5470b35dd5e221e4f090915b51b4fb2b36