Autogenerate Attach/Launch Dialogs [1/N]: First Version
Summary:
This is the first functional version of autogenerating launch and attach dialogs from a declarative configuration.

The `AutoGenLaunchAttachProvider` takes in a config of type `AutoGenConfig` and from that renders the `AutoGenLaunchAttachUiComponent`.

For now this replaces the Node Launch, Node Attach, and Python Launch launch attach providers and components. I will next expand this to consume all other launch attach providers and then remove any common boilerplate.

Reviewed By: mostafaeweda

Differential Revision: D7175025

fbshipit-source-id: 48ae2e15e1ded49b666f346e22a5e319938ee003