Store latest_diff in revisionStatusDisplay
Summary: We'll use this info to fetch the build statuses from sandcastle, who require this number and not the diff number. I'm storing it here because we have it for free, since it's already being requested in this same query. However the name RevisionStatusDisplay might be a little misleading now, since it's not just for displaying.

Reviewed By: tjfryan

Differential Revision: D7262361

fbshipit-source-id: 4585e5c2ba934c7bb886155129434b2145d8ad6e