Retry most long running hg observables
Summary:
In many places, we were assuming the observables we're watching would never throw. This is generally true, but in certain situations breakages occur, such as if a bad type gets sent across rpc.

In some situations in which a failure like this occurs, we can simply retry, and possibly fix the transient error. After a few retries, if we're still failing we should resort to throwing an error to the user so we can diagnose the root cause.

This should make smartlog more robust as a whole, for example if hg changes some message format in its internals.

The only risk here is if we retry observables that really shouldn't be retried, such as user-initiated actions. I don't expect this will be a problem since this new method is only used for long running observables that usually watch external events. Plus these callsites (almost) never throw exceptions.

Reviewed By: aadisriram

Differential Revision: D7281903

fbshipit-source-id: 38454c5ac9b45eb13548b02e5fce20e60772c056