Vsp Rewrite: Disposable Error Fix
Summary:
I'm not sure if this is the proper solution but if it isn't, then I'm not sure what the proper solution would be.

There was an error whenever you reloaded atom that said something like "Emitter has already been disposed". This is because the DebuggerLayoutManager needs to be disposed before the DebugService which when disposed, disposes the Model from DebuggerModel.js which disposes that emitter.

Reviewed By: mostafaeweda

Differential Revision: D7313189

fbshipit-source-id: 7f4c2913bde05771b129d7e34d9bb04cb7bc0ea0