send the client id in the header on a websocket connection to avoid the extra round-trip
Summary: Got to the airport a little early and thought I would see how hard this was. Turns out it was pretty straight-forward. I don't think I tested it enough, and it may still need some polish, but I figured I'd get it out before I got on the plane.

Reviewed By: hansonw

Differential Revision: D7053394

fbshipit-source-id: 3d3cad41463acfeb068798a8b0cdc238b12e1372