LSP hover can have null response
Summary:
Apparently textDocument/hover is supposed to return null when the LSP server has no hover to offer at that point. This diff makes Nuclide cope with that.

https://github.com/Microsoft/language-server-protocol/issues/261

(When I wrote this code originally, it wasn't clear what to return when the LSP server has no hover to offer. I surveyed existing LSP implementations and they returned either `[""]` or `[]`, so that's what Nuclide handled. But dbaumer clarified that the preferred return value is just `null`).

Reviewed By: hansonw

Differential Revision: D7351011

fbshipit-source-id: e1e6da2757619b0879bb6767bd1cddfc93e482fa