Java Debugger: Attach To Test [1/N]
Summary:
Video Example: https://pxl.cl/clNl

Clicking the debug button for the buck toolbar where the target is a test target will launch the tests with debugging enabled. The user then has to attach to the port specified in the console.

Next I'd like to make it so that the user doesn't need to manually attach to the debug port but since this was such a small diff I figured I'd put it up to get it out there.

Reviewed By: stepanhruda

Differential Revision: D7357083

fbshipit-source-id: 6e892282ac11291ec0cf005c250fb5ed6e6376a8