Removed providesModule tag from FB module format template used by several tools (e.g.: relay generator)
Summary:
This is required to remove the providesModule tag from many files in WWW. Most of them are generated by the Relay generator (~6K files).

There aren't any generated files with providesModule in fbsource, so we just need to commit the files with the templates.

In the diff that will be automatically created when this gets landed in fbsource, we should rebuild the code and commit any modified files so we don't have issues with hashes & CIGAR.

Reviewed By: mjesun

Differential Revision: D7370686

fbshipit-source-id: 46908c73c36204dc560df18b4cfd133f343dfaae