Make LspLanguageService fall back to language identifier regex when no hover range is given
Summary: Rather than falling back to a single character range when the remote LSP server doesn't return a range, we should use a simple word regex to make a range.  The hovered-over-identifier's position is used to calculate when to recheck for a new type hint, so this change will make it easier to mouse over the shown type hint without it going away.

Reviewed By: ljw1004

Differential Revision: D7390235

fbshipit-source-id: c8b769bbe0336f528f1613cff525a14b4ff7d070