Fix edge case with watchman events in filename search
Summary:
If a file is created and then rapidly modified, there's some buffering logic in PathSetUpdater that currently merges them into a single 'modified' event.

Only 'creation' events result in things getting added to the index, so this is wrong. We have to be careful to merge 'new' and 'modified' events into a 'new' event.

Reviewed By: wanderley

Differential Revision: D7397388

fbshipit-source-id: 2a4fbaf391f7ba660c114b8006c5531b55ffff80