Add Option to change font for nuclide-terminal
Summary:
I like to use the internal Terminal once in a while. But as xterm.js renders the terminal in a canvas I have to set the font in xterm.js and can't just change the css.
Sadly nuclide didn't have an option to change the font.
Now it does.
Closes https://github.com/facebook/nuclide/pull/1486

Reviewed By: hchau-fb

Differential Revision: D7400518

Pulled By: pelmers

fbshipit-source-id: aa2be3c5a8063fbea08e204282e444c5c026a35e